nokia/cbamlibrary/requirements.txt
requests==2.20.1
python-dotenv==0.10.3
